+++ src/components/ErrorNotification/index.js
@@ -1,0 +1,1 @@
+export * from './ErrorNotification';
