+++ src/types/OptionsType.ts
@@ -1,0 +1,12 @@
+export enum FilterName {
+  All = 'all',
+  Active = 'active',
+  Completed = 'completed',
+}
+
+export enum KeyValue {
+  Enter = 'Enter',
+  Esc = 'Escape',
+}
+
+export type OptionUpdate = 'all' | 'once';
